first whack at getting the konahdmi2rx to handle input video
Had to add some new paths to the code for the 4x style HDMI widgets that
the konahdmi2rx has.

This is a WIP and some extra debugging is left enabled. Next up will be
fix up the routing to use all the CSCs and LUTs so can do color
controls.