konahdmi2rx: The Brightness, Gamma and Gain controls now work.
Most of the work was in getting a working fixed point pow function and
the remainder was fighting with overflows in the fixed point math.

So far I have only tried the controls with stream 1, they should work
with stream 2 also, will test that once I get hue and saturation
working.